requirements.txt
python-telegram-bot
openai
pandas
requests
beautifulsoup4
